pyproject.toml
[tool.poetry]
name = "hermes"
package-mode = false
description = "A single pane of glass for static code analysis"
authors = ["Skelmis <skelmis.craft@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
piccolo = {extras = ["all"], version = "^1.15.0"}
litestar = {extras = ["cryptography"], version = "^2.10.0"}
uvicorn = "^0.30.3"
piccolo-admin = "^1.6.0"
python-dotenv = "^1.0.1"
orjson = "^3.10.6"
skelmis-commons = "^1.2.1"
apscheduler = {git = "https://github.com/agronholm/apscheduler.git"}


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
